+++ amd_openvx_extensions/amd_rpp/source/internal_publishKernels.cpp
@@ -162,6 +162,8 @@ vx_status get_kernels_to_publish()
     STATUS_ERROR_CHECK(ADD_KERNEL(Vignette_Register));
     STATUS_ERROR_CHECK(ADD_KERNEL(WarpAffine_Register));
     STATUS_ERROR_CHECK(ADD_KERNEL(Spectrogram_Register));
+    STATUS_ERROR_CHECK(ADD_KERNEL(Downmix_Register));
+    STATUS_ERROR_CHECK(ADD_KERNEL(ToDecibels_Register));
 
     return status;
 }