+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Documentation for MIVisionX is available at
 * Support for advanced GPUs
 * Support for PreEmphasis Filter augmentation in openVX extensions
 * Support for Spectrogram augmentation in openVX extensions
+* Support for Downmix and ToDecibels augmentations in openVX extensions
 
 ### Optimizations